refactor(direct-control): extract production choice source
Move the embedded production choice App UI source into packages/civ7-direct-control/src/play/operations/production-choice.ts while keeping the public wrapper/build command and production postcondition helpers in the facade.

Record the production-choice source owner and the ready-domain note disposition in the support modularization workstream. This is local source-relocation proof only, not runtime/live-game proof.

Validation: bun run --cwd packages/civ7-direct-control test; bun run --cwd packages/civ7-direct-control check; bun run --cwd packages/civ7-direct-control build; bun run openspec -- validate civ7-support-direct-control-modularization --strict; bun run resources:status; git diff --check.
